# museum/internal/infra/api/model_buy_museum_tickets_request.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Redocly Museum API
 *
 * An imaginary, but delightful Museum API for interacting with museum services and information. Built with love by Redocly.
 *
 * API version: 1.0.0
 * Contact: team@redocly.com
 */

package api




// BuyMuseumTicketsRequest - Request payload used for purchasing museum tickets.
type BuyMuseumTicketsRequest struct {

	TicketType TicketType `json:"ticketType"`

	// Identifier for a special event.
	EventId string `json:"eventId,omitempty"`

	TicketDate string `json:"ticketDate"`

	// Email address for ticket purchaser.
	Email string `json:"email"`

	// Phone number for the ticket purchaser (optional).
	Phone string `json:"phone,omitempty"`
}

// AssertBuyMuseumTicketsRequestRequired checks if the required fields are not zero-ed
func AssertBuyMuseumTicketsRequestRequired(obj BuyMuseumTicketsRequest) error {
	elements := map[string]interface{}{
		"ticketType": obj.TicketType,
		"ticketDate": obj.TicketDate,
		"email": obj.Email,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertBuyMuseumTicketsRequestConstraints checks if the values respects the defined constraints
func AssertBuyMuseumTicketsRequestConstraints(obj BuyMuseumTicketsRequest) error {
	return nil
}